pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "echo"
version = "1.0.1"
description = "Echo AI - Multi-agent AI orchestration system with plugin management"
authors = [
    { name = "Jonas Kahn", email = "me@ifelse.one" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13,<3.14"
keywords = ["ai", "agents", "echo-ai", "orchestration", "coordination", "langchain", "langgraph", "multi-agent", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
    "Environment :: Console",
    "Intended Audience :: System Administrators"
]
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "httpx>=0.28.1,<0.29.0",
    "langchain>=0.3.27,<0.4.0",
    "langchain-anthropic>=0.3.18,<0.4.0",
    "langchain-community>=0.3.27,<0.4.0",
    "langchain-core>=0.3.74,<0.4.0",
    "langchain-openai>=0.3.30,<0.4.0",
    "langgraph[add]>=0.6.5,<0.7.0",
    "langgraph-checkpoint-redis>=0.1.0,<0.2.0",
    "pydantic>=2.11.7,<3.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "redis>=6.4.0,<7.0.0",
    "uvicorn[standard]>=0.35.0,<0.36.0",
    "streamlit (>=1.48.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "echo-sdk (>=1.0.1,<2.0.0)",
]

[project.urls]
homepage = "https://github.com/jonaskahn/echo"
documentation = "https://echo.readthedocs.io/"
repository = "https://github.com/jonaskahn/echo.git"
issues = "https://github.com/jonaskahn/echo/issues"

[project.scripts]
echo-cli = "echo.cli:app"

[project.optional-dependencies]
dev = [
    "black>=25.1.0,<26.0.0",
    "isort>=6.0.1,<7.0.0",
    "mypy>=1.17.1,<2.0.0",
    "pre-commit>=4.3.0,<5.0.0",
    "pytest>=8.4.1,<9.0.0",
    "pytest-asyncio>=1.1.0,<2.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "pytest-mock>=3.14.1,<4.0.0",
    "ruff>=0.12.9,<0.13.0",
]

[tool.poetry]
packages = [{ include = "echo", from = "src" }]

# FOR LOCAL DEVELOPMENT
#[tool.poetry.group.local]
#optional = true
#
#[tool.poetry.group.local.dependencies]
#echo_sdk = {path = "./sdk", develop = true}
#echo_plugins = { path = "./plugins", develop = true }


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
linkchecker = ">=10.0.0"
mkdocs = ">=1.5.0"
mkdocs-git-revision-date-localized-plugin = ">=1.2.0"
mkdocs-material = ">=9.5.0"
mkdocs-mermaid2-plugin = ">=1.2.1"
mkdocs-minify-plugin = ">=0.7.0"
mkdocs-redirects = ">=1.2.0"
pymdown-extensions = ">=10.0.0"

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.13"
strict = true

[tool.ruff]
target-version = "py313"
line-length = 120